Moved LoginActivity to .login
