+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".LoginActivity"
+            android:name=".login.LoginActivity"
             android:configChanges="orientation|screenSize"
             android:launchMode="singleInstance">
             <intent-filter>
