+++ app/src/main/java/com/unbounds/trakt/login/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.unbounds.trakt;
+package com.unbounds.trakt.login;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -9,6 +9,9 @@ import android.webkit.WebChromeClient;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
+import com.unbounds.trakt.ApiWrapper;
+import com.unbounds.trakt.BuildConfig;
+import com.unbounds.trakt.R;
 import com.unbounds.trakt.api.HttpRequest;
 import com.unbounds.trakt.api.model.request.Code;
 import com.unbounds.trakt.api.model.response.Token;
