+++ app/src/main/res/layout/activity_login.xml
@@ -5,4 +5,4 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".LoginActivity" />
+    tools:context=".login.LoginActivity" />
